requirements.txt
Django==6.0.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Pillow==10.4.0
django-filter==24.2